Corrected logic for TeX generation in 'Scribe::generate(QuizType)' ..
Previously, the method tried to write the preview .tex files at the same
time as it was trying to write the answer-key. This can't be done.

Why? Because each preview needs its own \docbegin & \docend. And so,
if the preview is being written - and being written correctly - bu at the
	same time as the answer-key, then the answer-key will get multiple \docbegin
	& \docend - which is syntactically incorrect

Conversely, if the previews are written so that the answer-key is right, then
the previews would be wrong