writePreamble() should write only what is in preamble.tex ...
Previously, the method just had hard-coded println() calls within it.
And therefore, even if the preamble changed - which it has now to
include fig4tex - nothing changed in the generated .tex files. Need to
look out for anymore such hard-coding in Scribe

Also, something wrong with resolve() call on line 65. Was causing buildQuiz to barf ...