[Big one]: Rationalized build infrastructure to just two requests - writeTex and compileTex
The changes here work in conjunction with changes in gutenberg makefiles
and changes in the Rails app. These latter will be committed separately

If it all works, then we wouldn't need 3 separate requests - buildQuiz,
buildTestpaper and processWorksheet. Java need only write the bare
minimum information for Make to take over